backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
cors==1.0.1
aiohttp==3.9.1
pydantic==2.5.0
